refactor: moved enum to entity package
